template/test/requirements.txt
--index-url https://pypi.tuna.tsinghua.edu.cn/simple/
--extra-index-url https://mirrors.aliyun.com/pypi/simple/

locustio==1.0b2
